backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.0
redis==5.0.1
aiohttp==3.9.1
pytest==8.1.1
